+++ tiktool_core.py
@@ -33,7 +33,7 @@ OWNED_JOB_PREFIXES = ("restore-", "backup-", "transfer-")
 LICENSE_SECRET = base64.b64decode("J1+EGe8rOjhrROIE5Dygj2nnHTKH/9Bval0PMWI2w5E=")
 LICENSE_PREFIX = "IPTP-"
 ANSI_ESCAPE_RE = re.compile(r"\x1b(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])")
-APP_VERSION = "4.9.6"
+APP_VERSION = "4.9.8"
 LIBIMOBILE_VERSION = "1.4.0"